DUI：修复资源泄漏，修复BlurDrawDC坐标错误，修复元素渲染到混合位图时背景清除不正确，增加CCompositorPageAn；修复Dx资源泄漏；CEzD2D::MakeBitblt默认呈现方式改为DXGI_SWAP_EFFECT_SEQUENTIAL；增加MessageBoxTimeout定义

+++ eck/CMemWalker.h
@@ -316,7 +316,7 @@ struct CMemWalker
 	template<class T, class U, class V>
 	EckInline CMemWalker& operator>>(CRefStrT<T, U, V>& x)
 	{
-		const int cch = TcsLen((const T*)Data());
+		const int cch = (int)TcsLen((const T*)Data());
 		x.ReSize(cch);
 		return Read(x.Data(), (cch + 1) * sizeof(T));
 	}
